Frontend improvement for global map (#476)
* Bump version

* Add loading on the legend

* Update geometry request data to worker

* Improve query request for geometries

* Remove label when config changed

* Improve filtered geometries

* Fix moving fetching list of geography units to worker

* Update centroid create geometries

* Fix based on failed tests

* Add IS_TEST env for E2E tests

* Add delay to tests

* Increase delay to tests

+++ django_project/frontend/src/utils/WebWorker.jsx
@@ -20,11 +20,17 @@ export default class WebWorker {
   }
 }
 
-export function ExecuteWebWorker(script, data, callback) {
+export function ExecuteWebWorker(script, data, callback, onePostMessage = true) {
   var _webWorker = new WebWorker(script);
   _webWorker.postMessage(data);
   _webWorker.addEventListener('message', (event) => {
+    if (event.data?.isDone) {
+      _webWorker.terminate()
+      return
+    }
     callback(event.data)
-    _webWorker.terminate()
+    if (onePostMessage) {
+      _webWorker.terminate()
+    }
   });
 }
